+++ src/js/components/submissionsTable/DashboardContent.jsx
