+++ src/js/components/submissionsTable/filters/FilterBar.jsx
