+++ src/js/redux/actions/submissionsTable/appliedFilterActions.js
