+++ src/js/redux/reducers/submissionsTable/appliedFiltersReducer.js
